Refactor the paste role
Here we create a specific `PasteTextNode`, and
move most of the logic for formatting etc out of the `PasteNodesToDocutils` transform.

This makes the transform more generic, so that we can add more PasteNode subclasses to it in the future.

We also rename the role from `nb:paste` to `nb:text`,
to reflect its specific nature,
and rewrite the documentation to reflect this.