Allow Bokeh plots to be glued and pasted
Using the JSON-type output from Bokeh, this creates appropriate HTML and
Javascript nodes in the output when the Bokeh+JSON mimetype is found.
The Python Bokeh library is only imported when it is needed, to avoid a
package dependency on Bokeh. Users will need to install Bokeh via pip or
conda to be able to glue/paste Bokeh plots. Using the JSON output allows
plots to be pasted into figures as well as a general element with no
further changes.

Uses the Sphinx configuration to decide whether a page has Bokeh content
and only loads BokehJS on those pages. This simplifies the script that
needs to be embedded in each paste and ensures that Bokeh is only loaded
once from the CDN, unless the user runs output_notebook(). In the case
where the user runs output_notebook() the BokehJS library is loaded
twice but this doesn't appear to cause any problems.